pkg/extractor/options_test.go
/*
 * Copyright (c) 2018-2020 vChain, Inc. All Rights Reserved.
 * This software is released under GPL3.
 * The full license information can be found under:
 * https://www.gnu.org/licenses/gpl-3.0.en.html
 *
 */

package extractor

import (
	"errors"
	"testing"

	"github.com/stretchr/testify/assert"
)

type dummyOpts struct {
	flag bool
}

func withTestOption() Option {
	return func(o interface{}) error {
		if oo, ok := o.(*dummyOpts); ok {
			oo.flag = true
		}
		return nil
	}
}

func withTestOptionWithError() Option {
	return func(o interface{}) error {
		return errors.New("some error")
	}
}

func TestApply(t *testing.T) {

	opts := &dummyOpts{}

	err := Options([]Option{
		withTestOption(),
	}).Apply(opts)
	assert.NoError(t, err)
	assert.True(t, opts.flag)

	err = Options([]Option{
		withTestOptionWithError(),
	}).Apply(opts)
	assert.Error(t, err)
}
